### 2025-09-13 7238606
Geopolrisk FastAPI-API, Dockerfile, and related instructions
