### 2025-12-05 #8689 @mengqi-z
👋 @Haitham-ghaida @romainsacchi @nicolito - Would any of you be interested and available to review this submission for the Journal of Open Source Software (JOSS)? JOSS uses a checklist-driven review process, which you can find here: https://joss.readthedocs.io/en/latest/review_criteria.html
