### 2025-12-09 #9600 @editorialbot
[REVIEW]: geopolrisk-py: A Python-Based Library to Operationalize the Geopolitical Supply Risk Method for use in Life Cycle Assessment and Comparative Risk Assessment
**Submitting author:** @akoyamp (<a href="http://orcid.org/0000-0003-0839-0552">Anish Koyamparambath</a>)
**Repository:** https://github.com/akoyamp/geopolrisk-py
**Branch with paper.md** (empty if default branch):
**Version:** v2.0.0
