### 2025-12-10 #9600 @konstantinstadler
Installation guidance missing https://github.com/akoyamp/geopolrisk-py/issues/18
