### 2026-01-01 #9600 @FATelarico
### Software paper

> Summary (for a diverse, non-specialist audience)

The purpose and high-level functionality are stated clearly: the software “operationalize[s] the [GeoPolRisk] method” and “simplif[ies] the calculation process” by processing inputs (materials/countries/years) and producing outputs (scores/CFs). But the summary leans heavily on LCA/criticality terminology (“characterization factors”, “midpoint”, “comparative risk assessment”) without short, plain-language glosses. So, it is not that accessible yo a non-specialist audience.

> Statement of need (section titled “Statement of Need” + problem + audience + relation to other work) & State of the field

The paper is strong on the methodological context, weaker on software landscape. Practically, it does not clearly articulate software novelty relative to existing implementations/tools.

>  Quality of writing

Needs minor to moderate editing.

Concrete issues visible in the text:
- Placeholder / broken metadata: DOI “10.xxxxxx/draft”, “Published: unpublished”, “¿VOL?(¿ISSUE?)”, and an in-text note “<- update this…”.
- Typos/orthography: “refered” appears (and similar minor slips).
- Formatting artefacts: odd hyphenation/line-break issues (“Sus tainability”), equation formatting and spacing inconsistencies

> Bibliography

Some entries look incomplete or oddly specified (e.g., one reference line that reads like an internal identifier rather than a normal bibliographic entry).

### Suggestions

- One short sentence that explains `GeoPolRisk` in plain terms (e.g., “it estimates how likely supply disruptions are due to political instability and concentrated production”) and one sentence explicitly stating who benefits (e.g., “LCA practitioners and risk analysts in industry/public sector”).
- A short paragraph naming the closest alternatives (even if they are spreadsheets, proprietary tools, or ad hoc scripts), and stating what geopolrisk-py offers that they don’t (automation, reproducibility, built-in datasets/database, company-specific workflows, etc.).
- Double check the text for typos, misformatting
- Go over the references for artefacts.
