+++ paper.md
@@ -3,7 +3,7 @@ title: "geopolrisk-py: A Python-Based Library to Operationalize the Geopolitical
 tags:
   - Criticality
   - LCA
-  - Import
+  - Imports
   - Raw material
   - Market concentration
   - Abiotic Resources
@@ -28,19 +28,19 @@ affiliations:
     index: 2
 date: 18 July 2025
 bibliography: paper.bib
-aas-doi: 10.3847/xxxxx <- update this with the DOI from AAS once you know it.
+aas-doi: 10.3847/xxxxx
 aas-journal: The Journal of Open Source Software.
 ---
 
 # Summary
 
-The Geopolitical Supply Risk (GeoPolRisk) method assesses raw material criticality in Life Cycle Assessment (LCA), complementing traditional resource and environmental impact indicators. It is also applied for comparative risk assessment. However, calculating values for the GeoPolRisk method, such as characterization factors for the Geopolitical Supply Risk indicator and the supply risk score for comparative assessment, can be complex. To address this, the `geopolrisk-py` library has been developed to operationalize the method, simplifying the calculation process. This library processes data inputs like raw material names, countries, and years, making it more accessible. A notable feature is its ability to serve as a company-specific supply risk assessment tool. Future development will include integrating an uncertainty model and additional modules for complete compatibility with Brightway 2.5, further enhancing functionality and usability.
+The Geopolitical Supply Risk (GeoPolRisk) method estimates the likelihood that supplies of a given raw material may be disrupted due to political instability in producing countries and high concentration of global production. It is used in life cycle assessment (LCA) and in comparative risk assessments to complement environmental indicators with a supply-risk perspective. In practice, applying the method can be technically demanding, as it requires combining trade data, production statistics, governance indicators, and price information to compute supply-risk scores and characterization factors. The `geopolrisk-py` library operationalizes the GeoPolRisk method by automating these calculations in a transparent and reproducible manner. It takes structured inputs such as raw material names, countries or economic units, and reference years, and produces  supply-risk scores and characterization factors (CF)s for LCA. The library is intended for life cycle assessment practitioners, materials analysts, and risk analysts in industry and the public sector, and supports applications ranging from national and regional assessments to company-specific supply risk analyses. Future developments include the integration of uncertainty analysis and tighter coupling with the Brightway 2.5 framework to further enhance functionality and usability.
+
 
 # The GeoPolRisk Method
+The GeoPolRisk method was developed as an import-based indicator [@Gemechu2015] to integrate the supply risk of raw materials into LCA, proposed by [@Sonnemann2015]. It is designed to evaluate the supply risk from the perspective of an economic unit (country, trade block, region, group of countries, or company/organization) during a specific period. Since its inception, the method has evolved to incorporate multiple components of resource criticality [@SantillanSaldivar2022].
 
-The GeoPolRisk method is an import-based indicator [@Gemechu2015] developed to integrate the criticality of raw materials into the Area of Protection ‘natural resources’ in LCA, as proposed by [@Sonnemann2015]. It is designed to evaluate the supply risk of importing raw materials from the perspective of an economic unit (country, trade block, region, group of countries, or company/organization) during a specific period. Since its inception, the method has evolved to incorporate multiple components of resource criticality [@Cimprich2019].
-
[…]
