+++ paper.bib
@@ -1,215 +1,213 @@
 @article{Ciroth2016,
-author = {Ciroth, Andreas and Muller, St{\'{e}}phanie and Weidema, Bo and Lesage, Pascal},
-doi = {10.1007/s11367-013-0670-5},
-journal = {International Journal of Life Cycle Assessment},
-title = {{Empirically based uncertainty factors for the pedigree matrix in ecoinvent}},
-year = {2016}
+  author  = {Ciroth, Andreas and Muller, St{\'{e}}phanie and Weidema, Bo and Lesage, Pascal},
+  title   = {Empirically based uncertainty factors for the pedigree matrix in ecoinvent},
+  journal = {International Journal of Life Cycle Assessment},
+  year    = {2016},
+  doi     = {10.1007/s11367-013-0670-5}
 }
-@article{Rhoades1993TheHI,
-author = {Rhoades, Stephen A},
-journal = {Federal Reserve Bulletin},
-pages = {188--189},
-title = {{The Herfindahl-Hirschman index}},
-url = {https://api.semanticscholar.org/CorpusID:153018440},
-year = {1993}
+
+@article{Rhoades1993,
+  author  = {Rhoades, Stephen A},
+  title   = {The Herfindahl-Hirschman index},
+  journal = {Federal Reserve Bulletin},
+  year    = {1993},
+  pages   = {188--189},
+  url     = {https://api.semanticscholar.org/CorpusID:153018440}
 }
-@misc{cepii_baci_2024,
-annote = {BACI is distributed under the Etalab Open Licence 2.0, which means that any use is authorized, provided the source is mentioned. Mentioning BACI and the CEPII is an appropriate reference, ideally with a link to the webpage.},
-author = {CEPII},
-institution = {CEPII},
-title = {{BACI: International Trade Database at the Product-Level}},
-url = {https://www.cepii.fr/cepii/en/bdd_modele/bdd_modele_item.asp?id=37},
-urldate = {2024-09-20},
-year = {2024}
+
+@misc{Cepii2024,
+  author      = {CEPII},
+  title       = {BACI: International Trade Database at the Product-Level},
+  institution = {CEPII},
+  year        = {2024},
+  url         = {https://www.cepii.fr/cepii/en/bdd_modele/bdd_modele_item.asp?id=37},
+  urldate     = {2024-09-20}
 }
+
 @article{Cimprich2019,
-author = {Cimprich, Alexander and Bach, Vanessa and Helbig, Christoph and Thorenz, Andrea and Schrijvers, Dieuwertje and Sonnemann, Guido and Young, Steven B. and Sonderegger, Thomas and Berger, Markus},
-doi = {10.1111/jiec.12865},
-journal = {Journal of Industrial Ecology},
-keywords = {critical raw materials,industrial ecology,life cycle assessment,life cycle sustainability assessment,raw material criticality assessment,supply risk},
-number = {5},
-pages = {1226--1236},
-title = {{Raw material criticality assessment as a complement to environmental life cycle assessment: Examining methods for product-level supply risk assessment}},
-volume = {23},
-year = {2019}
+  author  = {Cimprich, Alexander and Bach, Vanessa and Helbig, Christoph and Thorenz, Andrea and Schrijvers, Dieuwertje and Sonnemann, Guido and Young, Steven B. and Sonderegger, Thomas and Berger, Markus},
+  title   = {Raw material criticality assessment as a complement to environmental life cycle assessment: Examining methods for product-level supply risk assessment},
+  journal = {Journal of Industrial Ecology},
+  year    = {2019},
+  volume  = {23},
+  number  = {5},
+  pages   = {1226--1236},
[…]
