+++ tests/__int__.py
@@ -1,16 +1,14 @@
-# Copyright (C) 2023 University of Bordeaux, CyVi Group & Anish Koyamparambath
+# Copyright (C) 2024 University of Bordeaux, CyVi Group & University of Bayreuth,
+# Ecological Resource Technology & Anish Koyamparambath, Christoph Helbig, Thomas Schraml
 # This file is part of geopolrisk-py library.
-#
 # geopolrisk-py is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
 # the Free Software Foundation, either version 3 of the License, or
 # (at your option) any later version.
-#
 # geopolrisk-py is distributed in the hope that it will be useful,
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
 # MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 # GNU General Public License for more details.
-#
 # You should have received a copy of the GNU General Public License
 # along with geopolrisk-py.  If not, see <https://www.gnu.org/licenses/>.
 
