+++ requirements.txt
@@ -2,3 +2,4 @@ pandas
 tqdm
 geopy
 openpyxl
+pytest

[…]
