+++ paper.bib
@@ -142,7 +142,7 @@
   url       = {http://dx.doi.org/10.1016/j.jclepro.2016.07.214}
 }
 
-@article{Bachmann2021,
+@report{Bachmann2021,
   author  = {Bachmann, Till M and Hackenhaar, Isadora C. and Horn, Rafael and Charter, Martin and Gehring, Florian and Graf, Roberta and Huysveld, Sophie and Alvarenga, Rodrigo A. F.},
   title   = {Critical evaluation of material criticality and product-related circularity approaches},
   year    = {2021},
