+++ paper.md
@@ -28,17 +28,15 @@ affiliations:
     index: 2
 date: 18 July 2025
 bibliography: paper.bib
-aas-doi: 10.3847/xxxxx
-aas-journal: The Journal of Open Source Software.
 ---
 
 # Summary
 
-The Geopolitical Supply Risk (GeoPolRisk) method estimates the likelihood that supplies of a given raw material may be disrupted due to political instability in producing countries and high concentration of global production. It is used in life cycle assessment (LCA) and in comparative risk assessments to complement environmental indicators with a supply-risk perspective. In practice, applying the method can be technically demanding, as it requires combining trade data, production statistics, governance indicators, and price information to compute supply-risk scores and characterization factors. The `geopolrisk-py` library operationalizes the GeoPolRisk method by automating these calculations in a transparent and reproducible manner. It takes structured inputs such as raw material names, countries or economic units, and reference years, and produces  supply-risk scores and characterization factors (CF)s for LCA. The library is intended for life cycle assessment practitioners, materials analysts, and risk analysts in industry and the public sector, and supports applications ranging from national and regional assessments to company-specific supply risk analyses. Future developments include the integration of uncertainty analysis and tighter coupling with the Brightway 2.5 framework to further enhance functionality and usability.
+The Geopolitical Supply Risk (GeoPolRisk) method estimates the likelihood that supplies of a given raw material may be disrupted due to political instability in producing countries and high concentration of global production. It is used in life cycle assessment (LCA) and in comparative risk assessments to complement environmental indicators with a supply-risk perspective. In practice, applying the method can be technically demanding, as it requires combining trade data, production statistics, governance indicators, and price information to compute supply risk scores and characterization factors. The `geopolrisk-py` library operationalizes the GeoPolRisk method by automating these calculations in a transparent and reproducible manner. It takes structured inputs such as raw material names, countries or economic units, and reference years, and produces  supply risk scores and characterization factors (CF)s for LCA. The library is intended for life cycle assessment practitioners, materials analysts, and risk analysts in industry and the public sector, and supports applications ranging from national and regional assessments to company specific supply risk analyses. Future developments include the integration of uncertainty analysis and tighter coupling with the Brightway 2.5 framework to further enhance functionality and usability.
 
 
 # The GeoPolRisk Method
-The GeoPolRisk method was developed as an import-based indicator [@Gemechu2015] to integrate the supply risk of raw materials into LCA, proposed by [@Sonnemann2015]. It is designed to evaluate the supply risk from the perspective of an economic unit (country, trade block, region, group of countries, or company/organization) during a specific period. Since its inception, the method has evolved to incorporate multiple components of resource criticality [@SantillanSaldivar2022].
+The GeoPolRisk method was developed as an import based indicator [@Gemechu2015] to integrate the supply risk of raw materials into LCA, proposed by [@Sonnemann2015]. It is designed to evaluate the supply risk from the perspective of an economic unit (country, trade block, region, group of countries, or company/organization) during a specific period. Since its inception, the method has evolved to incorporate multiple components of resource criticality [@SantillanSaldivar2022].
 
 The core of the method is the GeoPolRisk Score, which theoretically represents the probability of supply disruption due to geopolitical factors for a given raw material “A” and economic unit “c” during a specific year. As shown in Equation 1, this score is composed of two main components: (i) the global production concentration, represented by the Herfindahl-Hirschman Index (HHI) [@Rhoades1993] $HHI_A$, and (ii) the import dependence of the economic unit or also referred to as "import risk", which accounts for how much of the raw material is imported from politically unstable sources.
 
@@ -48,7 +46,7 @@ $$
[…]
