+++ tests/README.md
@@ -1,75 +1,157 @@
 # Test Suite for geopolrisk-py
 
-This directory contains the automated test suite for **geopolrisk-py**.
-The tests are designed to verify the correctness, internal consistency, and
-stability of the GeoPolRisk implementation.
+This directory contains the automated test suite for **geopolrisk-py**. The tests are designed to verify the correctness, internal consistency, and long-term stability of the GeoPolRisk implementation.
 
-The test suite includes unit tests for core computational functions as well as
-integration and regression tests that validate the full calculation workflow
-against reference results.
+The suite includes unit tests for core computational components, as well as integration and regression tests that validate the full calculation workflow against established reference results.
 
-## Test structure
 
-The test suite is organised by functional scope:
+## Test scope and structure
 
-- **`test_core.py`**  
-  Tests the numerical core of the GeoPolRisk method, including HHI, import risk,
-  GeoPolRisk scores, and characterization factor calculations.
+The test suite is organised by functional scope as follows:
 
-- **`test_database.py`**  
-  Tests low-level database utilities, ensuring correct execution of SQLite
-  queries used internally by the library.
+* **`test_core.py`**
+  Validates the numerical core of the GeoPolRisk method, including HHI calculations, import risk metrics, GeoPolRisk scores, and characterization factor computation.
 
-- **`test_util.py`**  
-  Tests supporting utility functions for data handling, country conversion,
-  aggregation, and result formatting.
+* **`test_database.py`**
+  Tests low-level database utilities, ensuring correct execution of internal SQLite queries and database interactions.
 
-- **`test_main.py`**  
-  Tests the main user-facing workflow (`gprs_calc`), verifying that complete
-  calculations return results with the expected structure for both country-level
-  and regional assessments.
+* **`test_util.py`**
+  Covers supporting utility functions related to data handling, country and region conversion, aggregation logic, and result formatting.
 
-- **`test_geopolrisk_data_results_from_excelfile.py`**  
-  Provides regression tests that compare computed results against reference
-  values extracted from a tool-based Excel test case. These tests ensure that
-  end-to-end results remain consistent with a known reference implementation.
+* **`test_main.py`**
+  Tests the main user-facing workflow (`gprs_calc`), verifying that complete calculations return results with the expected structure for both country-level and regional assessments.
 
----
+* **`test_geopolrisk_data_results_from_excelfile.py`**
+  Provides regression tests that compare computed results against reference values derived from an Excel-based test case. These tests ensure that end-to-end results remain consistent with a known reference implementation.
+
 
 ## Running the tests
 
-Tests are not included in the PyPI distribution and must be run from a cloned
-copy of the repository.
+Tests are not included in the PyPI distribution and must be executed from a cloned copy of the repository.
 
-Clone the repository locally:
+### 1. Clone the repository
 
 ```bash
 git clone https://github.com/akoyamp/geopolrisk-py.git
 cd geopolrisk-py
 ```
 
-From the repository root:
+## 2. Create a Python environment (Python 3.11 required)
+
+**Python 3.11 is mandatory.**
+Always specify the version explicitly when creating the environment.
+
+Choose **one** of the following methods.
+
+---
+
+### Option A — `venv` (recommended)
+
+Requires Python 3.11 to be installed on the system.
+
+```bash
+# Windows
+py -3.11 -m venv venv
+
+# macOS / Linux
+python3.11 -m venv venv
+```
+
+---
+
+### Option B — Conda (Anaconda / Miniconda)
+
+Recommended for users already working in Conda-based workflows.
+
+```bash
[…]
