### 2026-01-23 #9600 @akoyamp
Hi @mengqi-z ,

Thank you very much for following up, and for handling the editorial process.

I would also like to thank @FATelarico  and @konstantinstadler for their detailed and constructive comments. I fully agree with the feedback provided.

I have addressed all comments related to the software documentation, in particular by clarifying the installation instructions, testing instructions, and contribution guidance. These updates have been pushed to the GitHub repository.

One important point is that the library is not yet published on PyPI, although installation instructions currently describe the PyPI workflow. My intention is to upload the final, accepted version to PyPI as a stable version 2 release at the end of the review process.
If this is not acceptable, I can proceed with uploading the current version to PyPI immediately.

Regarding the manuscript itself, I have implemented the suggested revisions by @FATelarico  to improve clarity and to better emphasise the software novelty. I am currently waiting for final feedback from my co-authors.

Please let me know if you would like me to proceed differently at this stage.

Best regards,
Anish
