### 2026-01-26 #9600 @FATelarico
@mengqi-z I can confirm that my review is complete and I _do_ recommend acceptance based on the  [pre-2026](https://joss.readthedocs.io/en/pre-2026-submission/) editorial guidelines.
