### 2026-01-27 #9600 @mengqi-z
@FATelarico - Awesome! Thank you so much for taking the time to review the JOSS paper!
