### 2026-01-30 #9600 @konstantinstadler
@akoyamp : thanks for the installation part. However, following the installation guide I can not run the example or the notebook from the documentation.

```
In [1]: from geopolrisk.assessment.utils import default_rmlist
   ...: rawmaterials_list = default_rmlist()
The file doesn't exist. Please enter a folder path:
The following tables are missing from the database: ['commodityHS', 'Country_ISO', 'Aluminium', 'Antimony', 'Asbestos', 'Barytes', 'Bismuth', 'Cadmium', 'Chromium', 'Coal', 'Cobalt', 'Copper', 'Gold', 'Graphite', 'Iron', 'Lead', 'Lithium', 'Magnesite', 'Magnesium', 'Manganese', 'Mercury', 'Molybdenum', 'Nickel', 'Crude_Oil', 'REE', 'Silver', 'Tin', 'Titanium', 'Tungsten', 'Uranium', 'Zinc', 'Zirconium', 'NG', 'WGI', 'Price']
Could not extract the tables Execution failed on sql 'SELECT * FROM commodityHS': no such table: commodityHS
```

However, cloning the source code and installing in edit mode, this part works. I assume, you do not include the db in the normally installed version. Please clearify that, perhaps add a smaller example dataset or have a separate version with the db included.

In any case, after installing the package (python 3.11) and trying the example notebook it breaks at cell 4:

```
gprs_calc(
    ListofYear, ListofCountries, ListofMetals
)  # This is the main function that calculates the GeoPolRisk values for the specified list of countries, years and metals
# The function returns a pandas DataFrame with the results, which is also saved as an excel file in a folder named "ouput" in the document folder of the users home directory

Calculating the GeoPolRisk: 100%|██████████████████████████████████████████████████████████████| 212/212 [00:03<00:00, 69.38it/s]
--- Logging error ---
Traceback (most recent call last):
  File "/home/konstans/tmp/geopolrisk-py/src/geopolrisk/assessment/main.py", line 111, in gprs_calc
    writetodb(result)
  File "/home/konstans/tmp/geopolrisk-py/src/geopolrisk/assessment/utils.py", line 126, in writetodb
    exists = execute_query(check_query, db_path=dbpath, params=(row["DBID"],))
             ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/home/konstans/tmp/geopolrisk-py/src/geopolrisk/assessment/database.py", line 49, in execute_query
    cursor.execute(query, params or [])
sqlite3.OperationalError: no such column: DBID

Traceback (most recent call last):
  File "/home/konstans/.local/share/uv/python/cpython-3.11.13-linux-x86_64-gnu/lib/python3.11/logging/__init__.py", line 1110, in emit
    msg = self.format(record)
          ^^^^^^^^^^^^^^^^^^^
  File "/home/konstans/.local/share/uv/python/cpython-3.11.13-linux-x86_64-gnu/lib/python3.11/logging/__init__.py", line 953, in format
    return fmt.format(record)
           ^^^^^^^^^^^^^^^^^^
  File "/home/konstans/.local/share/uv/python/cpython-3.11.13-linux-x86_64-gnu/lib/python3.11/logging/__init__.py", line 687, in format
    record.message = record.getMessage()
                     ^^^^^^^^^^^^^^^^^^^
  File "/home/konstans/.local/share/uv/python/cpython-3.11.13-linux-x86_64-gnu/lib/python3.11/logging/__init__.py", line 377, in getMessage
    msg = msg % self.args
          ~~~~^~~~~~~~~~~
TypeError: not all arguments converted during string formatting
Call stack:
  File "<frozen runpy>", line 198, in _run_module_as_main
  File "<frozen runpy>", line 88, in _run_code
  File "/home/konstans/tmp/geopolrisk-py/.venv/lib/python3.11/site-packages/ipykernel_launcher.py", line 18, in <module>
    app.launch_new_instance()
  File "/home/konstans/tmp/geopolrisk-py/.venv/lib/python3.11/site-packages/traitlets/config/application.py", line 1075, in launch_instance
    app.start()
  File "/home/konstans/tmp/geopolrisk-py/.venv/lib/python3.11/site-packages/ipykernel/kernelapp.py", line 758, in start
    self.io_loop.start()
  File "/home/konstans/tmp/geopolrisk-py/.venv/lib/python3.11/site-packages/tornado/platform/asyncio.py", line 211, in start
    self.asyncio_loop.run_forever()
  File "/home/konstans/.local/share/uv/python/cpython-3.11.13-linux-x86_64-gnu/lib/python3.11/asyncio/base_events.py", line 608, in run_forever
    self._run_once()
  File "/home/konstans/.local/share/uv/python/cpython-3.11.13-linux-x86_64-gnu/lib/python3.11/asyncio/base_events.py", line 1936, in _run_once
    handle._run()
  File "/home/konstans/.local/share/uv/python/cpython-3.11.13-linux-x86_64-gnu/lib/python3.11/asyncio/events.py", line 84, in _run
    self._context.run(self._callback, *self._args)
  File "/home/konstans/tmp/geopolrisk-py/.venv/lib/python3.11/site-packages/ipykernel/kernelbase.py", line 614, in shell_main
    await self.dispatch_shell(msg, subshell_id=subshell_id)
  File "/home/konstans/tmp/geopolrisk-py/.venv/lib/python3.11/site-packages/ipykernel/kernelbase.py", line 471, in dispatch_shell
    await result
  File "/home/konstans/tmp/geopolrisk-py/.venv/lib/python3.11/site-packages/ipykernel/ipkernel.py", line 366, in execute_request
    await super().execute_request(stream, ident, parent)
  File "/home/konstans/tmp/geopolrisk-py/.venv/lib/python3.11/site-packages/ipykernel/kernelbase.py", line 827, in execute_request
    reply_content = await reply_content
  File "/home/konstans/tmp/geopolrisk-py/.venv/lib/python3.11/site-packages/ipykernel/ipkernel.py", line 458, in do_execute
    res = shell.run_cell(
  File "/home/konstans/tmp/geopolrisk-py/.venv/lib/python3.11/site-packages/ipykernel/zmqshell.py", line 663, in run_cell
    return super().run_cell(*args, **kwargs)
  File "/home/konstans/tmp/geopolrisk-py/.venv/lib/python3.11/site-packages/IPython/core/interactiveshell.py", line 3123, in run_cell
    result = self._run_cell(
  File "/home/konstans/tmp/geopolrisk-py/.venv/lib/python3.11/site-packages/IPython/core/interactiveshell.py", line 3178, in _run_cell
    result = runner(coro)
  File "/home/konstans/tmp/geopolrisk-py/.venv/lib/python3.11/site-packages/IPython/core/async_helpers.py", line 128, in _pseudo_sync_runner
    coro.send(None)
  File "/home/konstans/tmp/geopolrisk-py/.venv/lib/python3.11/site-packages/IPython/core/interactiveshell.py", line 3400, in run_cell_async
    has_raised = await self.run_ast_nodes(code_ast.body, cell_name,
  File "/home/konstans/tmp/geopolrisk-py/.venv/lib/python3.11/site-packages/IPython/core/interactiveshell.py", line 3641, in run_ast_nodes
    if await self.run_code(code, result, async_=asy):
  File "/home/konstans/tmp/geopolrisk-py/.venv/lib/python3.11/site-packages/IPython/core/interactiveshell.py", line 3701, in run_code
    exec(code_obj, self.user_global_ns, self.user_ns)
  File "/tmp/ipykernel_2642804/3260329966.py", line 1, in <module>
    gprs_calc(
  File "/home/konstans/tmp/geopolrisk-py/src/geopolrisk/assessment/main.py", line 114, in gprs_calc
    logging.debug("Error while recording data into dataframe", e)
Message: 'Error while recording data into dataframe'
Arguments: (OperationalError('no such column: DBID'),)
```
