### 2026-01-30 #9600 @konstantinstadler
@akoyamp : one other thing: In the online documentation you write:
"The traded price is calculated using the “free on board value,” a mechanism commonly used in international trade that considers the cost, insurance, and freight of a product being transported from the seller to the buyer. "

However, your definition is for cif, not fob - please clearify
