### 2026-01-30 #9600 @konstantinstadler
@akoyamp : The paper is well written, and good examples.
However, I miss the "State of the field: Do the authors describe how this software compares to other commonly-used packages?"

One more point: as at least in the testing setup 3rd party data is included, I think you need to specify which license the data has and, based on that,  under which license the data processed by the package can be used.
