### 2026-01-30 #9600 @konstantinstadler
@akoyamp  regarding testing: the review asks for automated tests. I see you already use github CI - can you add the tests to the CI or, if not, why do you choose not too?
