+++ paper.md
@@ -32,7 +32,7 @@ bibliography: paper.bib
 
 # Summary
 
-The Geopolitical Supply Risk (GeoPolRisk) method estimates the likelihood that supplies of a given raw material may be disrupted due to political instability in producing countries and high concentration of global production. It is used in life cycle assessment (LCA) and in comparative risk assessments to complement environmental indicators with a supply-risk perspective. In practice, applying the method can be technically demanding, as it requires combining trade data, production statistics, governance indicators, and price information to compute supply risk scores and characterization factors. The `geopolrisk-py` library operationalizes the GeoPolRisk method by automating these calculations in a transparent and reproducible manner. It takes structured inputs such as raw material names, countries or economic units, and reference years, and produces  supply risk scores and characterization factors (CF)s for LCA. The library is intended for life cycle assessment practitioners, materials analysts, and risk analysts in industry and the public sector, and supports applications ranging from national and regional assessments to company specific supply risk analyses. Future developments include the integration of uncertainty analysis and tighter coupling with the Brightway 2.5 framework to further enhance functionality and usability.
+The Geopolitical Supply Risk (GeoPolRisk) method estimates the likelihood that supplies of a given raw material may be disrupted due to political instability in producing countries and high concentration of global production. It is used in life cycle assessment (LCA) and in comparative risk assessments to complement environmental indicators with a supply-risk perspective. In practice, applying the method can be technically demanding, as it requires combining trade data, production statistics, governance indicators, and price information to compute supply risk scores and characterization factors. The `geopolrisk-py` library operationalizes the GeoPolRisk method by automating these calculations in a transparent and reproducible manner. It takes structured inputs such as raw material names, countries or economic units, and reference years, and produces  supply risk scores and characterization factors (CF)s for LCA. The library is intended for life cycle assessment practitioners, materials scientists, and risk analysts in industry and the public sector, and supports applications ranging from national and regional assessments to company specific supply risk analyses. Future developments include the integration of uncertainty analysis and tighter coupling with the Brightway 2.5 framework to further enhance functionality and usability.
 
 
 # The GeoPolRisk Method
@@ -66,7 +66,7 @@ $$
 
 # Statement of Need
 
-The GeoPolRisk method has been recommended by several international initiatives concerned with the integration of criticality aspects into life cycle assessment. These include the Global Guidance for Life Cycle Impact Assessment Indicators and Methods, coordinated by the United Nations Environment Programme [@Berger2020], and the ORIENTING project [@Hackenhaar2022], a European research initiative that critically reviewed methods for integrating Criticality Assessment within a life cycle sustainability assessment framework. The method is also recognized by the International Round Table on Materials Criticality [@Schrijvers2020] as a relevant approach for evaluating short-term supply disruptions of raw materials in a geopolitical context.
+The GeoPolRisk method has been recommended by several international initiatives concerned with the best practice of resource use in life cycle assessment. These include the Global Guidance for Life Cycle Impact Assessment Indicators and Methods, coordinated by the United Nations Environment Programme [@Berger2020], and the ORIENTING project [@Hackenhaar2022], a European research initiative that critically reviewed methods for integrating Criticality Assessment within a life cycle sustainability assessment framework. The method is also recognized by the International Round Table on Materials Criticality [@Schrijvers2020] as a relevant approach for evaluating short-term supply disruptions of raw materials in a geopolitical context.
 
[…]
