### 2026-02-11 b920d66
Update - Paper (Licensing Clarity)
