+++ docs/conf.py
@@ -24,6 +24,8 @@ nb_execution_mode = "off"
 # -- Options for HTML output -------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output
 
-html_theme = "alabaster"
+html_theme = "sphinx_rtd_theme"
 html_static_path = ["_static"]
 html_css_files = ["custom.css"]
+html_logo = "_static/GPRS.png"
+html_favicon = "_static/GPRS.ico"
