+++ docs/conf.py
@@ -27,5 +27,5 @@ nb_execution_mode = "off"
 html_theme = "sphinx_rtd_theme"
 html_static_path = ["_static"]
 html_css_files = ["custom.css"]
-html_logo = "_static/GPRS.png"
-html_favicon = "_static/GPRS.ico"
+html_logo = "_static/logo.png"
+html_favicon = "_static/icon.ico"
