### 2026-02-11 #9600 @akoyamp
@konstantinstadler, thank you very much for your feedback and comments on the code.

1. With regard to the installation of the package: I have tested it again in a Linux environment and it works. However, `pip install geopolrisk-py` will not work at this stage because, although the README specifies this method, I intend to upload only the final accepted version to PyPI. If this is not acceptable, I can upload the current version to PyPI immediately, which will resolve the issue. In the meantime, a functional beta version is available via: `pip install geopolrisk-py==2.0.0b1`.
2. With regard to the library databases in the `lib` folder: they are intended to be included for both testing and regular use. I have reviewed the code again and fixed the related issues. It is working now in regular usage and also in testing.
3. Concerning automated testing, I have now configured GitHub CI to run the tests, and it works as expected.
4. Regarding the documentation: you are absolutely right. The value is calculated using FOB, and I have corrected the explanation accordingly.
5. Regarding the “state of the field”: as recommended, I had already included a comparison in the “Statement of need” (in lines 72-82) describing how the method is currently implemented. Is this sufficient?
6. Regarding the licensing of the background database: this is a valid point. I have now clarified the licensing of both the code and the background datasets in the README and in the documentation.
