### 2026-02-13 #9600 @konstantinstadler
@mengqi-z : I can also recommend the article for acceptance.
I put some further issues on the code examples on the issue tracker of the source repo, but I guess this is outside the scope of the JOSS review.

@akoyamp : Many thanks for the answers and for making this important package open source! Hope to see it used in Brightway and other packages soon!
