### 2026-02-13 #9600 @mengqi-z
@akoyamp - Thank you for your quick responses to the reviewers! We are very close to get your paper accepted. However, the errors identified in the examples should be resolved, or the steps required to prevent the errors described in issue (https://github.com/akoyamp/geopolrisk-py/issues/18) should be clearly documented in the example instructions and confirmed by the reviewer prior to acceptance.

@konstantinstadler - Thank you for taking the time and effort to review this. As noted above, could you please confirm that the errors have been resolved in the issue https://github.com/akoyamp/geopolrisk-py/issues/18? Once that is completed, I can move forward with next steps for acceptance. Thank you.
