### 2026-02-13 #9600 @akoyamp
@konstantinstadler  @mengqi-z  I have updated the documentation accordingly. The installation section now includes guidance on removing older local databases before upgrading, and a reminder has been added in the usage section. For the company-level assessment, I clarified how to reproduce the example correctly.
Thank you both for the helpful feedback.
