### 2026-02-20 #9600 @mengqi-z
@akoyamp - I've proofread the paper and all looks good.

After this point could you:

- [ ] Double check authors and affiliations (including ORCIDs).
- [ ] Make a release of the software with the latest changes from the review and **post the version number here**. This is the version that will be used in the JOSS paper.
- [ ] Archive the release on Zenodo/figshare/etc and **post the DOI here**.
