### 2026-02-20 #9600 @akoyamp
@konstantinstadler Thank you again for reviewing the work. May I now close the issue opened in my repository?

@mengqi-z As requested:

> * [ ] Double check authors and affiliations (including ORCIDs).

Authors, affiliations, and ORCIDs have been carefully verified and are correct.

> * [ ] Make a release of the software with the latest changes from the review and **post the version number here**. This is the version that will be used in the JOSS paper.

A release has been created incorporating the latest changes. The version to be used in the JOSS paper is v2.0.0

> * [ ] Archive the release on Zenodo/figshare/etc and **post the DOI here**.

The release has been archived on Zenodo. The DOI is: https://doi.org/10.5281/zenodo.18713988

> * [ ] Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.

The title and author list, including ORCIDs, match exactly those in the JOSS paper.

> * [ ] Make sure that the license listed for the archive is the same as the software license.

The license listed in the archive is the same as the software license.
