### 2026-02-20 akoyamp/geopolrisk-py#21 @mengqi-z
Fix JOSS Paper Issue
Hi @akoyamp ,

It seems we are running into a small issue when compiling the JOSS paper during the acceptance checks: https://github.com/openjournals/joss-reviews/issues/9600#issuecomment-3936099659

I looked through the paper and believe that this line below in `paper.md` might be where the issue sits:

> ..., yielding the GeoPolRisk Midpoint [@SantillanSaldivar2022](Equation 2)

Could you please try change this line to
> ..., yielding the GeoPolRisk Midpoint (Equation 2) [@SantillanSaldivar2022]

Then we can try this again. Thanks!

openjournals/joss-reviews#9600
