+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "geopolrisk-py"
-version = "2.0.0.beta2"
+version = "2.0.0"
 description = "Python library for geopolitical supply risk assessment (GeoPolRisk)"
 readme = "README.md"
 requires-python = ">=3.10,<3.12"
@@ -14,7 +14,7 @@ authors = [
 ]
 keywords = ["LCA", "criticality", "geopolitical", "supply-risk", "raw-materials"]
 classifiers = [
-  "Development Status :: 4 - Beta",
+  "Development Status :: 5 - Production/Stable",
   "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
   "Programming Language :: Python :: 3",
   "Programming Language :: Python :: 3.10",
