### 2026-02-20 akoyamp/geopolrisk-py#21 @mengqi-z
Thank you @akoyamp ! Everything works great now!
