### 2026-02-20 #9600 @kyleniemeyer
Hi @akoyamp, your paper needs a few minor edits before we accept:
- on the first page, please edit the citation command in "proposed by (Sonnemann et al., 2015)." to use the `@author:year` style rather than `[@author:2001]`, to format it as "proposed by Sonnemann et al. (2015)."
- It looks like some details are missing from the Bachmann et al. citation, it isn't clear what this is or where one could find it.
- The FMRA citation just points to the ministry home page; I assume you intended to cite a specific database / dataset?
- The Mutel citation is missing the journal name.
- Rhoades citation: the article title needs some `{` and `}` to protect capitalization (this may be an issue in other titles)
- the Sacchi et al. citation is missing volume, issue, page numbers (may apply to other recent publications cited)
- The URL in the World Bank citation is broken
