+++ paper.bib
@@ -1,21 +1,21 @@
-@article{Ciroth2016,
+@article{Ciroth:2016,
   author  = {Ciroth, Andreas and Muller, St{\'{e}}phanie and Weidema, Bo and Lesage, Pascal},
-  title   = {Empirically based uncertainty factors for the pedigree matrix in ecoinvent},
+  title   = {Empirically Based Uncertainty Factors for the Pedigree Matrix in Ecoinvent},
   journal = {International Journal of Life Cycle Assessment},
   year    = {2016},
   doi     = {10.1007/s11367-013-0670-5}
 }
 
-@article{Rhoades1993,
+@article{Rhoades:1993,
   author  = {Rhoades, Stephen A},
-  title   = {The Herfindahl-Hirschman index},
+  title   = {The {Herfindahl-Hirschman Index}},
   journal = {Federal Reserve Bulletin},
   year    = {1993},
   pages   = {188--189},
   url     = {https://api.semanticscholar.org/CorpusID:153018440}
 }
 
-@misc{Cepii2024,
+@misc{Cepii:2024,
   author      = {CEPII},
   title       = {BACI: International Trade Database at the Product-Level},
   institution = {CEPII},
@@ -24,9 +24,9 @@
   urldate     = {2024-09-20}
 }
 
-@article{Cimprich2019,
+@article{Cimprich:2019,
   author  = {Cimprich, Alexander and Bach, Vanessa and Helbig, Christoph and Thorenz, Andrea and Schrijvers, Dieuwertje and Sonnemann, Guido and Young, Steven B. and Sonderegger, Thomas and Berger, Markus},
-  title   = {Raw material criticality assessment as a complement to environmental life cycle assessment: Examining methods for product-level supply risk assessment},
+  title   = {Raw Material Criticality Assessment as a Complement to Environmental Life Cycle Assessment: Examining Methods for Product-level Supply Risk Assessment},
   journal = {Journal of Industrial Ecology},
   year    = {2019},
   volume  = {23},
@@ -35,9 +35,9 @@
   doi     = {10.1111/jiec.12865}
 }
 
-@article{Hackenhaar2022,
+@article{Hackenhaar:2022,
   author    = {Hackenhaar, Isadora and Alvarenga, Rodrigo A.F. and Bachmann, Till M. and Riva, Federico and Horn, Rafael and Graf, Roberta and Dewulf, Jo},
-  title     = {A critical review of criticality methods for a European Life Cycle Sustainability Assessment},
+  title     = {A Critical Review of Criticality Methods for a European Life Cycle Sustainability Assessment},
   journal   = {Procedia CIRP},
   year      = {2022},
   volume    = {105},
@@ -48,9 +48,9 @@
   url       = {https://doi.org/10.1016/j.procir.2022.02.071}
 }
 
-@article{SantillanSaldivar2022,
+@article{SantillanSaldivar:2022,
   author    = {Santill{\'{a}}n-Saldivar, Jair and Gemechu, Eskinder and Muller, St{\'{e}}phanie and Villeneuve, Jacques and Young, Steven B. and Sonnemann, Guido},
-  title     = {An improved resource midpoint characterization method for supply risk of resources: integrated assessment of Li-ion batteries},
+  title     = {An Improved Resource Midpoint Characterization Method for Supply Risk of Resources: Integrated Assessment of Li-ion Batteries},
   journal   = {The International Journal of Life Cycle Assessment},
   year      = {2022},
   volume    = {27},
@@ -61,9 +61,10 @@
   url       = {https://doi.org/10.1007/s11367-022-02027-y}
 }
 
-@article{Mutel2017,
+@article{Mutel:2017,
   author = {Mutel, Chris},
[…]
