+++ paper.md
@@ -36,9 +36,9 @@ The Geopolitical Supply Risk (GeoPolRisk) method estimates the likelihood that s
 
 
 # The GeoPolRisk Method
-The GeoPolRisk method was developed as an import based indicator [@Gemechu2015] to integrate the supply risk of raw materials into LCA, proposed by [@Sonnemann2015]. It is designed to evaluate the supply risk from the perspective of an economic unit (country, trade block, region, group of countries, or company/organization) during a specific period. Since its inception, the method has evolved to incorporate multiple components of resource criticality [@SantillanSaldivar2022].
+The GeoPolRisk method was developed as an import based indicator [@Gemechu:2015] to integrate the supply risk of raw materials into LCA, proposed by @Sonnemann:2015. It is designed to evaluate the supply risk from the perspective of an economic unit (country, trade block, region, group of countries, or company/organization) during a specific period. Since its inception, the method has evolved to incorporate multiple components of resource criticality [@SantillanSaldivar:2022].
 
-The core of the method is the GeoPolRisk Score, which theoretically represents the probability of supply disruption due to geopolitical factors for a given raw material “A” and economic unit “c” during a specific year. As shown in Equation 1, this score is composed of two main components: (i) the global production concentration, represented by the Herfindahl-Hirschman Index (HHI) [@Rhoades1993] $HHI_A$, and (ii) the import dependence of the economic unit or also referred to as "import risk", which accounts for how much of the raw material is imported from politically unstable sources.
+The core of the method is the GeoPolRisk Score, which theoretically represents the probability of supply disruption due to geopolitical factors for a given raw material “A” and economic unit “c” during a specific year. As shown in Equation 1, this score is composed of two main components: (i) the global production concentration, represented by the Herfindahl-Hirschman Index (HHI) [@Rhoades:1993] $HHI_A$, and (ii) the import dependence of the economic unit or also referred to as "import risk", which accounts for how much of the raw material is imported from politically unstable sources.
 
 Equation 1:
 
@@ -48,7 +48,7 @@ $$
 
 Here, $g_i$ is the political (in)stability of exporter $i$, $f_{Aic}$ is the amount of raw material $A$ imported by $c$ from $i$, $p_{Ac}$ is the domestic production, and $F_{Ac}$ is the total imports.
 
-To integrate this method into life cycle impact assessment by associating the GeoPolRisk score with mass of the raw material, a factor is developed by multiplying the score with the market price of the material, yielding the GeoPolRisk Midpoint (Equation 2) [@SantillanSaldivar2022]. This value represents the potential value of raw material at imminent risk per kilogram of raw material consumed.
+To integrate this method into life cycle impact assessment by associating the GeoPolRisk score with mass of the raw material, a factor is developed by multiplying the score with the market price of the material, yielding the GeoPolRisk Midpoint (Equation 2) [@SantillanSaldivar:2022]. This value represents the potential value of raw material at imminent risk per kilogram of raw material consumed.
 
 Equation 2:
 
@@ -56,7 +56,7 @@ $$
 \text{GeoPolRisk\_Midpoint}_{Ac} = HHI_A \cdot \sum_i \left( \frac{g_i \cdot f_{Aic}}{p_{Ac} + F_{Ac}} \right) \cdot \bar{p}
 $$
 
-$\bar{p}$ is the price of the raw material. To enable comparison between materials and product systems, the midpoint factor is normalized using copper as a reference. The value for copper is calculated as an average of all the countries and for 5 years. This yields the Geopolitical Supply Risk Potential (GSP) [@Koyamparambath2024], as shown in Equation 3. The GSP has units of kg Cu-eq/kg$_A$ and is referred to as the CF for the Geopolitical Supply Risk indicator.
[…]
