### 2026-02-20 1b02d24
Update - Paper (Fix: Citations - Title case)
