+++ paper.bib
@@ -1,14 +1,18 @@
 @article{Ciroth:2016,
   author  = {Ciroth, Andreas and Muller, St{\'{e}}phanie and Weidema, Bo and Lesage, Pascal},
-  title   = {Empirically Based Uncertainty Factors for the Pedigree Matrix in Ecoinvent},
+  title   = {{Empirically Based Uncertainty Factors for the Pedigree Matrix in Ecoinvent}},
   journal = {International Journal of Life Cycle Assessment},
   year    = {2016},
-  doi     = {10.1007/s11367-013-0670-5}
+  volume  = {21},
+  number  = {9},
+  pages   = {1338--1349},
+  doi     = {10.1007/s11367-013-0670-5},
+  url     = {https://doi.org/10.1007/s11367-013-0670-5}
 }
 
 @article{Rhoades:1993,
   author  = {Rhoades, Stephen A},
-  title   = {The {Herfindahl-Hirschman Index}},
+  title   = {{The Herfindahl-Hirschman Index}},
   journal = {Federal Reserve Bulletin},
   year    = {1993},
   pages   = {188--189},
@@ -17,7 +21,7 @@
 
 @misc{Cepii:2024,
   author      = {CEPII},
-  title       = {BACI: International Trade Database at the Product-Level},
+  title       = {{BACI: International Trade Database at the Product-Level}},
   institution = {CEPII},
   year        = {2024},
   url         = {https://www.cepii.fr/cepii/en/bdd_modele/bdd_modele_item.asp?id=37},
@@ -26,7 +30,7 @@
 
 @article{Cimprich:2019,
   author  = {Cimprich, Alexander and Bach, Vanessa and Helbig, Christoph and Thorenz, Andrea and Schrijvers, Dieuwertje and Sonnemann, Guido and Young, Steven B. and Sonderegger, Thomas and Berger, Markus},
-  title   = {Raw Material Criticality Assessment as a Complement to Environmental Life Cycle Assessment: Examining Methods for Product-level Supply Risk Assessment},
+  title   = {{Raw Material Criticality Assessment as a Complement to Environmental Life Cycle Assessment: Examining Methods for Product-level Supply Risk Assessment}},
   journal = {Journal of Industrial Ecology},
   year    = {2019},
   volume  = {23},
@@ -37,7 +41,7 @@
 
 @article{Hackenhaar:2022,
   author    = {Hackenhaar, Isadora and Alvarenga, Rodrigo A.F. and Bachmann, Till M. and Riva, Federico and Horn, Rafael and Graf, Roberta and Dewulf, Jo},
-  title     = {A Critical Review of Criticality Methods for a European Life Cycle Sustainability Assessment},
+  title     = {{A Critical Review of Criticality Methods for a European Life Cycle Sustainability Assessment}},
   journal   = {Procedia CIRP},
   year      = {2022},
   volume    = {105},
@@ -50,7 +54,7 @@
 
 @article{SantillanSaldivar:2022,
   author    = {Santill{\'{a}}n-Saldivar, Jair and Gemechu, Eskinder and Muller, St{\'{e}}phanie and Villeneuve, Jacques and Young, Steven B. and Sonnemann, Guido},
-  title     = {An Improved Resource Midpoint Characterization Method for Supply Risk of Resources: Integrated Assessment of Li-ion Batteries},
+  title     = {{An Improved Resource Midpoint Characterization Method for Supply Risk of Resources: Integrated Assessment of Li-ion Batteries}},
   journal   = {The International Journal of Life Cycle Assessment},
   year      = {2022},
   volume    = {27},
@@ -63,7 +67,7 @@
 
 @article{Mutel:2017,
   author = {Mutel, Chris},
[…]
