### 2026-02-20 #9600 @akoyamp
@kyleniemeyer  thank you for the review. I have addressed all comments related to citations in line with your suggestions and corrected a few additional minor inconsistencies as well.
Please let me know if any further changes are needed.
