### 2026-02-20 #9600 @kyleniemeyer
@akoyamp one final thing: it looks like you need `{` `}` around "World Bank" in the author name; BibTeX is currently interpreting that as a person's name ("Bank, W.").
