+++ paper.bib
@@ -118,7 +118,7 @@
 }
 
 @misc{WorldBank:2024,
-  author  = {World Bank},
+  author  = {{World Bank}},
   title   = {{Worldwide Governance Indicators}},
   year    = {2024},
   url     = {https://www.govindicators.org/},
