+++ paper.bib
@@ -87,7 +87,7 @@
 
 @article{Koyamparambath:2021,
   author  = {Koyamparambath, Anish and Santillan-Saldivar, Jair and Mclellan, Benjamin and Sonnemann, Guido},
-  title   = {{Supply Risk Evolution of Raw Materials for Batteries and Fossil Fuels for Selected OECD Countries ( 2000 - 2018 )}},
+  title   = {{Supply Risk Evolution of Raw Materials for Batteries and Fossil Fuels for Selected OECD Countries (2000 - 2018)}},
   journal = {Resources Policy},
   year    = {2021},
   volume  = {75},
